requirements.txt
pandas==2.2.3
matplotlib==3.9.4
akshare==1.16.42
chinese_calendar==1.10.0
jieba==0.42.1
wordcloud==1.9.4
zhdate==0.1
beautifulsoup4==4.12.3
ttrss-python==0.5
requests==2.32.3
lxml==5.3.1
Jinja2==3.1.5
PyYAML==6.0.2
tqdm==4.67.1
numpy==2.0.2
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.13.1
statsmodels==0.14.4
Flask==3.0.3
Werkzeug==3.0.6
MarkupSafe==3.0.2
idna==3.10
certifi==2025.1.31
urllib3==2.3.0
charset-normalizer==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.16.0
setuptools==75.8.0
pip==24.2
wheel==0.44.0